created config.Markets types.go file
